Wallpaper.py

+++ plugins/wall.py
@@ -1,29 +1,35 @@
+"""Wallpaper Module"""
 from userge import userge, Message, Config
 from bs4 import BeautifulSoup as soup
 import requests
 from random import randint, choice
 import os
+import asyncio
+from PIL import Image
 
-CHANNEL = userge.getCLogger(__name__)
+down_p = str(Config.DOWN_PATH.rstrip('/'))
 
 
 async def dlimg(link):
     e = requests.get(link).content
     paea = 'donno.{}'.format(link.split('.')[-1])
-    path_i = os.path.join(Config.DOWN_PATH, paea)
+    path_i = os.path.join(down_p, paea)
     with open(path_i, 'wb') as k:
         k.write(e)
     return path_i
 
 
-async def wall(strin: str):
+async def walld(strin: str):
     if len(strin.split()) > 1:
         strin = '+'.join(strin.split())
     url = 'https://wall.alphacoders.com/search.php?search='
-    none_got = 'https://wall.alphacoders.com/finding_wallpapers.php'
+    none_got = [
+        'https://wall.alphacoders.com/finding_wallpapers.php',
+        'https://wall.alphacoders.com/search-no-results.php'
+        ]
     page_link = 'https://wall.alphacoders.com/search.php?search={}&page={}'
     resp = requests.get(f'{url}{strin}')
-    if resp.url == none_got:
+    if resp.url in none_got:
         return False
     if 'by_category.php' in resp.url:
         page_link = str(resp.url).replace('&amp;', '') + '&page={}'
@@ -31,13 +37,6 @@ async def wall(strin: str):
     else:
         check_link = False
     resp = soup(resp.content, 'lxml')
-    wall_num = resp.find('h1', {'class': 'center title'})
-    wall_num = list(wall_num.text.split())
-    for i in wall_num:
-        try:
-            wall_num = int(i)
-        except ValueError:
-            pass
     try:
         page_num = resp.find('div', {'class': 'visible-xs'})
         page_num = page_num.find('input', {'class': 'form-control'})
@@ -73,7 +72,10 @@ async def wall(strin: str):
     except Exception:
         pass
     del list_a_s
-    tit_link = choice(tit_links)
+    if len(tit_links) != 0:
+        tit_link = choice(tit_links)
+    else:
+        return False
     return tit_link
 
 
@@ -81,40 +83,47 @@ async def wall(strin: str):
     'header': "Search Wallpaper",
     'description': '''Search and Download Hd Wallpaper from
 alphacoders and upload to Telegram''',
-    'usage': "{tr}wall [Query]",
-    'examples': "{tr}wall luffy"
+    'usage': "{tr}wall [Query], [Second Query]",
+    'examples': "{tr}wall luffy",
+    'Another Example': "`{tr}wall Luffy, Naruto`"
 })
 async def idk_sir(message: Message):
-    if not os.path.isdir(Config.DOWN_PATH):
-        os.makedirs(Config.DOWN_PATH)
+    if not os.path.isdir(down_p):
+        os.makedirs(down_p)
+    cat_id = message.chat.id
     if message.input_str:
         qu = message.input_str
-        await message.edit('**Processing...**')
-        await CHANNEL.log(f"Search Query : {qu}")
-        try:
-            link = await wall(str(qu))
-        except Exception as e:
-            await message.edit(e)
-            return
-        if link:
-            idl = await dlimg(link[0])
-            await message.edit('**Uploading...**')
-            if not len(link[1].split()) < 11:
-                capo = '**' + ' '.join(link[1].split()[:11]) + '**'
-            else:
-                capo = '**' + link[1] + '**'
+        for q in qu.split(','):
+            q = q.strip()
+            await message.edit(f'**Processing...**\n**Searching for **`{q}`')
             try:
-                cat_id = message.chat.id
-                await userge.send_photo(cat_id, idl, caption=capo)
-                await userge.send_document(message.chat.id, idl)
+                link = await walld(str(q))
             except Exception as e:
                 await message.edit(e)
                 return
-            await message.delete()
-            os.remove(idl)
-        else:
-            await message.edit('**Result Not Found**')
-            await message.reply_sticker('CAADBAADIQEAAl_GARknbPJaYsVA2xYE')
+            if link:
+                idl = await dlimg(link[0])
+                if link[0].endswith('png'):
+                    im = Image.open(idl)
+                    idl = idl.replace('png', 'jpeg')
+                    im = im.convert('RGB')
+                    im.save(idl, 'jpeg')
+                    os.remove(idl)
+                await message.edit('**Uploading...**')
+                if not len(link[1].split()) < 11:
+                    capo = '**' + ' '.join(link[1].split()[:11]) + '**'
+                else:
+                    capo = '**' + link[1] + '**'
+                try:
+                    await userge.send_photo(cat_id, idl, caption=capo)
+                    await userge.send_document(cat_id, idl)
+                    os.remove(idl)
+                except Exception as e:
+                    await message.edit(e)
+            else:
+                await message.edit('**Result Not Found**')
+        await asyncio.sleep(3)
+        await message.delete()
     else:
         await message.edit('**Give me Something to search.**')
-        await message.reply_sticker('CAADAQADmQADTusQR6fPCVZ3EhDoFgQ')
+        await userge.send_sticker(cat_id, 'CAADAQADmQADTusQR6fPCVZ3EhDoFgQ')
